feat #6 : refactor User and Seller entities to member domain with BaseEntity inheritance
