+++ backend/src/main/java/com/phonebid/app/member/domain/Provider.java
@@ -1,4 +1,4 @@
-package com.phonebid.app.user.domain;
+package com.phonebid.app.member.domain;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
